Image Classification using CNN, Transfer Learning, Comparision to VGG16 and Inception V3 and Dashboard Design
 classification of images on two different datasets containing a
total of 8992 images of which the model used 75 percent dataset to train and 25 percent dataset
to test the results. The accuracy obtained by using a basic Sequential CNN model is almost 73
percent. However, the transfer learning model for fashion dataset is not as accurate as the
model for flowers dataset. This can be clearly understood by looking at the confusion matrices
in the below figure. The prediction results are not satisfactory for the hoodies-female and
sweatshirts.
The results generated by the pretrained models of Oxford (VGG16) and Google (Inception V3)
is more accurate by a difference of 10 percent than our custom model. We would conclude that
given the computation power and the time required for the model to generate the results,
Sequential CNN model is a good working model for the specific datasets. We are quite satisfied
with the results generated.
Although this model requires time to setup, the results generated are quite accurate and reliable
to the best of our knowledge. The accuracy of the custom model for both the datasets is neither
high nor low. This model is a good exercise for understanding the methods of image
classification the problems that arise when implementing such a model. After a few
improvisations and fine-tunings S-CNN can be used for other datasets and as an API for the
on-the-go format.